setup skeleton
