+++ front-end/src/App.css
@@ -1,2 +1,7 @@
-.App {
+.App { 
+    display: flex;
+    flex-direction: row;
+
+    width: 100%;
+	height: 100%;
 }
