+++ front-end/src/index.css
@@ -4,6 +4,10 @@ body {
 		"Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
+
+	background-color: #E5E5E5;
+	width: 100%;
+	height: 100%;
 }
 
 code {
