+++ front-end/src/organisms/Sidebar.js
@@ -1,18 +1,19 @@
 import React from "react";
 import LogoutButton from "../atoms/LogoutButton";
 import "./Sidebar.css";
-class Sidebar extends React.Component {
-  render() {
-    return (
-      <div className= "sidebar">
-        <b>Welcome, Emmanuel!</b>
-        <div>
-                    Please Select a Workspace...
-        </div>
-        <LogoutButton/>
-      </div>
-    );
-  }
+
+
+function Sidebar() {
+
+  return (
+    <div className="SideBar">
+      <div>logo</div>
+      <div className="SideBar-greeting">My Workspace</div>
+      <div>Manage Account</div>
+      <LogoutButton/>
+    </div>
+  );
+
 }
 
 export default Sidebar;
