+++ front-end/src/pages/Dashboard.css
@@ -1,0 +1,3 @@
+.Dashboard{
+    margin-left: 200px; /* size of sidebar */
+}
